# 001 - manual-build/config/api-docs.yaml
openapi: 3.0.1
info:
  title: FaCT API
  description: API for all operations relating to the FaCT application.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
...
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /courts:
    get:
      tags:
      - courts-controller
      summary: "Find courts by name, address, town or postcode"
      operationId: findCourtByNameOrAddressOrPostcodeOrTown
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtReference"
    post:
      tags:
      - admin-courts-controller
      summary: Add a new court
      operationId: addNewCourt
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCourt"
        required: true
      responses:
        "401":
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "409":
          description: Court already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
...
  /courts/{slug}:
    get:
      tags:
      - courts-controller
      summary: Find court details by slug
      operationId: findCourtByName
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
    delete:
      tags:
      - admin-courts-controller
      summary: Delete a court
      operationId: deleteCourt
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorised
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
...
  /courts/search:
    get:
      tags:
      - courts-controller
      summary: Return active courts based on a provided prefix
      operationId: getCourtsBySearch
      parameters:
      - name: prefix
        in: query
        required: true
        schema:
          maxLength: 1
          minLength: 1
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtReference"
...
  /courts/court-history/search:
    get:
      tags:
      - courts-controller
      summary: Return active court based a search of old court names
      operationId: getCourtByCourtHistoryNameSearch
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Court History Found but No corresponding Court found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtReferenceWithHistoricalName"
        "200":
          description: Successful - returns empty CourtReferenceWithHistoricalName
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtReferenceWithHistoricalName"
        "204":
          description: "Successful - but no history found, so returns no content"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtReferenceWithHistoricalName"
...
components:
  schemas:
    Court:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        name_cy:
          type: string
        info:
          type: string
        info_cy:
          type: string
        open:
          type: boolean
        in_person:
          type: boolean
        access_scheme:
          type: boolean
        opening_times:
          type: array
          items:
            $ref: "#/components/schemas/OpeningTime"
        common_platform:
          type: boolean
        urgent_message:
          type: string
        urgent_message_cy:
          type: string
        types:
          type: array
          items:
            $ref: "#/components/schemas/CourtType"
    CourtType:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        search:
          type: string
        code:
          type: integer
          format: int32
    OpeningTime:
      type: object
      properties:
        description:
          type: string
        hours:
          type: string
        description_cy:
          type: string
...
    NewCourt:
      type: object
      properties:
        new_court_name:
          maxLength: 200
          minLength: 1
          pattern: "^[a-zA-Z0-9-'() ]*$"
          type: string
        service_centre:
          type: boolean
        service_areas:
          type: array
          items:
            type: string
        lon:
          type: number
          format: double
        lat:
          type: number
          format: double
...
    CourtReference:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        updated_at:
          type: string
        displayed:
          type: boolean
        region:
          type: integer
          format: int32
...
    CourtReferenceWithHistoricalName:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        updated_at:
          type: string
        displayed:
          type: boolean
        region:
          type: integer
          format: int32
        historicalName:
          type: string
...